basic websocket connection stuff and made a setup.py
