+++ nsb_server/requirements.txt
@@ -6,3 +6,5 @@ Jinja2==2.11.2
 MarkupSafe==1.1.1
 Werkzeug==1.0.1
 firebase-admin
+Flask-SocketIO
+functools
